Move SendGhWorldtubeData callback to Cce lib
